Fix pet star export values for profession crafting and combat
We correctly render the table cells based on the new bonus values
for profession, crafting and combat bonus columns.

However, when exporting as CSV, the values are not correctly matched
and return NaN instead of the star count.  While this doesn't really
provide all of the data, let's at least make this portion of the
export work again.

Fix this by stripping off the bits higher than 0xFF and use a simple
range check to figure out what the star count is:
  0 = 0
  1-79 = 1
  80-159 = 2
  160+ = 3

In the future, we could add additional columns showing the
- profession quest skill (matching the egg type)
- profession bonus name and scalar value